pyproject.toml
[project]
name = "ai-ppt-generator"
version = "0.1.0"
description = "企业级 AI PPT 全自动生成系统"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.0",
    "litellm>=1.0.0",
    "markitdown[all]>=0.0.1",
    "httpx>=0.27.0",
    "python-multipart>=0.0.9",
    "python-pptx>=0.6.23",
    "pillow>=10.0.0",
    "websockets>=12.0",
"tavily-python>=0.5.0",
"graphviz>=0.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = ["pytest>=8.0", "pytest-asyncio>=0.24"]
